WIP: Move package key determination logic to `FlowPackageKey`
